backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.1",
    "openai>=2.29.0",
    "openai-whisper>=20250625",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "uvicorn[standard]>=0.42.0",
]
